requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.9.5
aioredis==1.3.1
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
apache-libcloud==3.8.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
bcrypt==4.2.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.8
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.2
fastapi-cache2==0.2.2
fasteners==0.19
filelock==3.16.1
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.7
httpx==0.28.1
identify==2.6.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==3.0.1
mccabe==0.7.0
multidict==6.1.0
mypy==1.12.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
phonenumbers==8.13.49
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.0
propcache==0.2.0
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.3
pyasn1==0.6.1
pycodestyle==2.12.1
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
pydantic_core==2.23.4
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
PyYAML==6.0.2
redis==4.6.0
requests==2.32.3
rsa==4.9
ruff==0.6.9
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sqladmin==0.19.0
SQLAlchemy==2.0.35
sqlalchemy-file==0.6.0
SQLAlchemy-Utils==0.41.2
starlette==0.39.2
time-machine==2.16.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.20.0
virtualenv==20.26.6
wheel==0.38.4
WTForms==3.1.2
yarl==1.15.1
dishka==1.4.2

orjson==3.10.15
aiojobs==1.3.0
aiormq==6.8.1
hyx==0.0.2
